@@ -6,9 +6,6 @@
   "dependencies": {
     "@emotion/react": "11.11.0",
     "@emotion/styled": "11.11.0",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.13.0",
     "@testing-library/jest-dom": "^5.11.4",
@@ -27,16 +24,14 @@
     "linkify-it": "^5.0.0",
     "linkify-react": "^4.2.0",
     "markdown-to-jsx": "^7.1.0",
-    "material-ui-color": "^1.2.0",
     "mic-recorder-to-mp3": "^2.2.2",
     "moment": "^2.29.4",
-    "mui-audio-player": "^1.0.1",
     "or": "^0.2.0",
     "qrcode.react": "^1.0.0",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-color": "^2.19.3",
     "react-csv": "^2.2.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-flatpickr": "^3.10.13",
     "react-gauge-chart": "^0.4.1",
     "react-modal-image": "^2.5.0",
@@ -48,7 +43,8 @@
     "recharts": "^2.0.2",
     "socket.io-client": "^4.7.3",
     "use-sound": "^2.0.1",
-    "yup": "^0.32.8"
+    "yup": "^0.32.8",
+    "@mui/styles": "^5.13.0"
   },
   "scripts": {
     "start": "react-scripts --openssl-legacy-provider start",
