@@ -1,4 +1,4 @@
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 
 import React, { useRef, useEffect, useState } from "react";
 
