@@ -1,26 +1,26 @@
 import React from "react";
 import * as Yup from "yup";
-import { makeStyles } from "@material-ui/core/styles";
-import Stepper from "@material-ui/core/Stepper";
-import Step from "@material-ui/core/Step";
-import StepLabel from "@material-ui/core/StepLabel";
-import StepContent from "@material-ui/core/StepContent";
+import makeStyles from '@mui/styles/makeStyles';
+import Stepper from "@mui/material/Stepper";
+import Step from "@mui/material/Step";
+import StepLabel from "@mui/material/StepLabel";
+import StepContent from "@mui/material/StepContent";
 import api from "../../services/api";
-import Typography from "@material-ui/core/Typography";
-import EditIcon from "@material-ui/icons/Edit";
-import { IconButton } from "@material-ui/core";
+import Typography from "@mui/material/Typography";
+import EditIcon from "@mui/icons-material/Edit";
+import { IconButton } from "@mui/material";
 import { Formik, Field, FieldArray } from "formik";
-import DeleteOutline from "@material-ui/icons/DeleteOutline";
-import SaveIcon from "@material-ui/icons/Save";
-import TextField from "@material-ui/core/TextField";
+import DeleteOutline from "@mui/icons-material/DeleteOutline";
+import SaveIcon from "@mui/icons-material/Save";
+import TextField from "@mui/material/TextField";
 import toastError from "../../errors/toastError";
 import { toast } from "react-toastify";
-import HelpOutlineOutlinedIcon from "@material-ui/icons/HelpOutlineOutlined";
+import HelpOutlineOutlinedIcon from "@mui/icons-material/HelpOutlineOutlined";
 import CustomToolTip from "../ToolTips";
 import ConfirmationModal from "../ConfirmationModal";
 import { i18n } from "../../translate/i18n";
-import Switch from "@material-ui/core/Switch";
-import { FormControlLabel } from "@material-ui/core";
+import Switch from "@mui/material/Switch";
+import { FormControlLabel } from "@mui/material";
 
 const QueueSchema = Yup.object().shape({
   options: Yup.array()
