@@ -1,30 +1,26 @@
 import React from "react";
-import Button from "@material-ui/core/Button";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import Typography from "@material-ui/core/Typography";
+import Button from "@mui/material/Button";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
+import Typography from "@mui/material/Typography";
 
 import { i18n } from "../../translate/i18n";
 
 const ConfirmationModal = ({ title, children, open, onClose, onConfirm }) => {
 	return (
-		<Dialog
+        <Dialog
 			open={open}
 			onClose={() => onClose(false)}
 			aria-labelledby="confirm-dialog"
 		>
-			<DialogTitle id="confirm-dialog">{title}</DialogTitle>
-			<DialogContent dividers>
+            <DialogTitle id="confirm-dialog">{title}</DialogTitle>
+            <DialogContent dividers>
 				<Typography>{children}</Typography>
 			</DialogContent>
-			<DialogActions>
-				<Button
-					variant="contained"
-					onClick={() => onClose(false)}
-					color="default"
-				>
+            <DialogActions>
+				<Button variant="contained" onClick={() => onClose(false)}>
 					{i18n.t("confirmationModal.buttons.cancel")}
 				</Button>
 				<Button
@@ -38,8 +34,8 @@ const ConfirmationModal = ({ title, children, open, onClose, onConfirm }) => {
 					{i18n.t("confirmationModal.buttons.confirm")}
 				</Button>
 			</DialogActions>
-		</Dialog>
-	);
+        </Dialog>
+    );
 };
 
 export default ConfirmationModal;
