@@ -1,15 +1,15 @@
 import React, { useState } from "react";
 
-import { makeStyles } from "@material-ui/core/styles";
-import Typography from "@material-ui/core/Typography";
-import IconButton from "@material-ui/core/IconButton";
-import CloseIcon from "@material-ui/icons/Close";
-import Drawer from "@material-ui/core/Drawer";
-import Link from "@material-ui/core/Link";
-import InputLabel from "@material-ui/core/InputLabel";
-//import Avatar from "@material-ui/core/Avatar";
-import Button from "@material-ui/core/Button";
-import Paper from "@material-ui/core/Paper";
+import makeStyles from '@mui/styles/makeStyles';
+import Typography from "@mui/material/Typography";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+import Drawer from "@mui/material/Drawer";
+import Link from "@mui/material/Link";
+import InputLabel from "@mui/material/InputLabel";
+//import Avatar from "@mui/material/Avatar";
+import Button from "@mui/material/Button";
+import Paper from "@mui/material/Paper";
 
 import { i18n } from "../../translate/i18n";
 
@@ -91,7 +91,7 @@ const ContactDrawer = ({ open, handleDrawerClose, contact, loading }) => {
 	const [modalOpen, setModalOpen] = useState(false);
 
 	return (
-		<Drawer
+        <Drawer
 			className={classes.drawer}
 			variant="persistent"
 			anchor="right"
@@ -106,17 +106,15 @@ const ContactDrawer = ({ open, handleDrawerClose, contact, loading }) => {
 				paper: classes.drawerPaper,
 			}}
 		>
-			<div className={classes.header}>
-				<IconButton
-					color="primary"
-					onClick={handleDrawerClose}>
+            <div className={classes.header}>
+				<IconButton color="primary" onClick={handleDrawerClose} size="large">
 					<CloseIcon />
 				</IconButton>
 				<Typography style={{ justifySelf: "center" }}>
 					{i18n.t("contactDrawer.header")}
 				</Typography>
 			</div>
-			{loading ? (
+            {loading ? (
 				<ContactDrawerSkeleton classes={classes} />
 			) : (
 				<div className={classes.content}>
@@ -173,8 +171,8 @@ const ContactDrawer = ({ open, handleDrawerClose, contact, loading }) => {
 					</Paper>
 				</div>
 			)}
-		</Drawer>
-	);
+        </Drawer>
+    );
 };
 
 export default ContactDrawer;
