@@ -1,16 +1,16 @@
 import React from "react";
-import Skeleton from "@material-ui/lab/Skeleton";
-import Typography from "@material-ui/core/Typography";
-import Paper from "@material-ui/core/Paper";
+import Skeleton from '@mui/material/Skeleton';
+import Typography from "@mui/material/Typography";
+import Paper from "@mui/material/Paper";
 import { i18n } from "../../translate/i18n";
 
 const ContactDrawerSkeleton = ({ classes }) => {
 	return (
-		<div className={classes.content}>
-			<Paper square variant="outlined" className={classes.contactHeader}>
+        <div className={classes.content}>
+            <Paper square variant="outlined" className={classes.contactHeader}>
 				<Skeleton
 					animation="wave"
-					variant="circle"
+					variant="circular"
 					width={160}
 					height={160}
 					className={classes.contactAvatar}
@@ -19,7 +19,7 @@ const ContactDrawerSkeleton = ({ classes }) => {
 				<Skeleton animation="wave" height={25} width={80} />
 				<Skeleton animation="wave" height={25} width={80} />
 			</Paper>
-			<Paper square className={classes.contactDetails}>
+            <Paper square className={classes.contactDetails}>
 				<Typography variant="subtitle1">
 					{i18n.t("contactDrawer.extraInfo")}
 				</Typography>
@@ -36,8 +36,8 @@ const ContactDrawerSkeleton = ({ classes }) => {
 					<Skeleton animation="wave" height={20} width={160} />
 				</Paper>
 			</Paper>
-		</div>
-	);
+        </div>
+    );
 };
 
 export default ContactDrawerSkeleton;
