@@ -1,10 +1,10 @@
 import React from "react";
 
-import { Avatar, Card, CardHeader, Typography } from "@material-ui/core";
-import Skeleton from "@material-ui/lab/Skeleton";
+import { Avatar, Card, CardHeader, Typography } from "@mui/material";
+import Skeleton from '@mui/material/Skeleton';
 
-import { makeStyles } from "@material-ui/core/styles";
-import { grey } from '@material-ui/core/colors';
+import makeStyles from '@mui/styles/makeStyles';
+import { grey } from '@mui/material/colors';
 
 const useStyles = makeStyles(theme => ({
 	cardAvatar: {
@@ -26,27 +26,24 @@ const useStyles = makeStyles(theme => ({
 export default function CardCounter(props) {
     const { icon, title, value, loading } = props
 	const classes = useStyles();
-    return ( !loading ? 
-        <Card>
-            <CardHeader
-                avatar={
-                    <Avatar className={classes.cardAvatar}>
-                        {icon}
-                    </Avatar>
-                }
-                title={
-                    <Typography variant="h6" component="h2" className={classes.cardTitle}>
-                        { title }
-                    </Typography>
-                }
-                subheader={
-                    <Typography variant="subtitle1" component="p" className={classes.cardSubtitle}>
-                        { value }
-                    </Typography>
-                }
-            />
-        </Card>
-        : <Skeleton variant="rect" height={80} />
-    )
+    return (!loading ? <Card>
+        <CardHeader
+            avatar={
+                <Avatar className={classes.cardAvatar}>
+                    {icon}
+                </Avatar>
+            }
+            title={
+                <Typography variant="h6" component="h2" className={classes.cardTitle}>
+                    { title }
+                </Typography>
+            }
+            subheader={
+                <Typography variant="subtitle1" component="p" className={classes.cardSubtitle}>
+                    { value }
+                </Typography>
+            }
+        />
+    </Card> : <Skeleton variant="rectangular" height={80} />);
     
 }
