@@ -1,7 +1,7 @@
 import React from "react";
 
-import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
+import makeStyles from '@mui/styles/makeStyles';
+import Container from "@mui/material/Container";
 
 const useStyles = makeStyles((theme) => ({
   mainContainer: {
