@@ -1,12 +1,12 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
+import makeStyles from '@mui/styles/makeStyles';
 import {
 	Tooltip,
 	Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 
 
-import { green } from "@material-ui/core/colors";
+import { green } from "@mui/material/colors";
 
 const useStyles = makeStyles(theme => ({
 	mainPaper: {
