@@ -3,8 +3,8 @@ import * as Yup from "yup";
 import { Formik, Form, Field } from "formik";
 import { toast } from "react-toastify";
 
-import { makeStyles } from "@material-ui/core/styles";
-import { green } from "@material-ui/core/colors";
+import makeStyles from '@mui/styles/makeStyles';
+import { green } from "@mui/material/colors";
 
 import {
   Dialog,
@@ -24,7 +24,7 @@ import {
   InputLabel,
   Select,
   Checkbox,
-} from "@material-ui/core";
+} from "@mui/material";
 
 import api from "../../services/api";
 import { i18n } from "../../translate/i18n";
