@@ -1,6 +1,6 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
+import makeStyles from '@mui/styles/makeStyles';
+import Container from "@mui/material/Container";
 import { i18n } from "../../translate/i18n";
 
 const useStyles = makeStyles(theme => ({
