@@ -2,8 +2,8 @@ import React, { useState, useCallback, useContext } from "react";
 import { toast } from "react-toastify";
 import { format, parseISO } from "date-fns";
 
-import { makeStyles } from "@material-ui/core/styles";
-import { green } from "@material-ui/core/colors";
+import makeStyles from '@mui/styles/makeStyles';
+import { green } from "@mui/material/colors";
 import {
 	Button,
 	TableBody,
@@ -16,7 +16,7 @@ import {
 	Tooltip,
 	Typography,
 	CircularProgress,
-} from "@material-ui/core";
+} from "@mui/material";
 import {
 	Edit,
 	CheckCircle,
@@ -27,7 +27,7 @@ import {
 	DeleteOutline,
 	SyncOutlined,
 	WhatsApp
-} from "@material-ui/icons";
+} from "@mui/icons-material";
 
 import MainContainer from "../../components/MainContainer";
 import MainHeader from "../../components/MainHeader";
@@ -199,8 +199,8 @@ const Connections = () => {
 
 	const renderActionButtons = whatsApp => {
 		return (
-			<>
-				{whatsApp.status === "qrcode" && (
+            <>
+                {whatsApp.status === "qrcode" && (
 					<Button
 						size="small"
 						variant="contained"
@@ -210,7 +210,7 @@ const Connections = () => {
 						{i18n.t("connections.buttons.qrcode")}
 					</Button>
 				)}
-				{whatsApp.status === "DISCONNECTED" && (
+                {whatsApp.status === "DISCONNECTED" && (
 					<>
 						<Button
 							size="small"
@@ -230,7 +230,7 @@ const Connections = () => {
 						</Button>
 					</>
 				)}
-				{(whatsApp.status === "CONNECTED" ||
+                {(whatsApp.status === "CONNECTED" ||
 					whatsApp.status === "PAIRING" ||
 					whatsApp.status === "TIMEOUT") && (
 						<Button
@@ -244,13 +244,13 @@ const Connections = () => {
 							{i18n.t("connections.buttons.disconnect")}
 						</Button>
 					)}
-				{whatsApp.status === "OPENING" && (
-					<Button size="small" variant="outlined" disabled color="default">
+                {whatsApp.status === "OPENING" && (
+					<Button size="small" variant="outlined" disabled>
 						{i18n.t("connections.buttons.connecting")}
 					</Button>
 				)}
-			</>
-		);
+            </>
+        );
 	};
 
 	const renderStatusToolTips = whatsApp => {
